9.1.269.40 source code

+++ build/apple/write_pkg_info.py
@@ -1,4 +1,4 @@
-# Copyright 2016 The Chromium Authors
+# Copyright 2016 The Chromium Authors. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
@@ -24,10 +24,8 @@ def Main():
   args = parser.parse_args()
 
   # Remove the output if it exists already.
-  try:
+  if os.path.exists(args.output):
     os.unlink(args.output)
-  except FileNotFoundError:
-    pass
 
   plist = plist_util.LoadPList(args.plist)
   package_type = plist['CFBundlePackageType']
@@ -49,4 +47,8 @@ def Main():
 
 
 if __name__ == '__main__':
+  # TODO(https://crbug.com/941669): Temporary workaround until all scripts use
+  # python3 by default.
+  if sys.version_info[0] < 3:
+    os.execvp('python3', ['python3'] + sys.argv)
   sys.exit(Main())
